Create controller api module
So that the cli code cannot import things outside the api
