+++ .dockerignore
@@ -1,0 +1,7 @@
+**
+!cli
+!services/controller
+**/*.egg-info
+**/__pycache__
+**/build
+**/dist
