--- cli/.dockerignore
@@ -1,4 +1,0 @@
-**/*.egg-info
-**/__pycache__
-build
-dist
