+++ cli/Dockerfile
@@ -1,17 +1,19 @@
 # Build
 FROM python:3-slim as builder
 
-RUN mkdir -p /src
-WORKDIR /src
+RUN mkdir -p /services/controller/src
+RUN mkdir -p /cli/src
+WORKDIR /cli
 
-COPY LICENSE MANIFEST.in README.rst setup.py ./
-COPY src src/
+COPY cli/LICENSE cli/MANIFEST.in cli/README.rst cli/setup.py ./
+COPY cli/src /cli/src/
+COPY services/controller/src /services/controller/src/
 RUN python setup.py bdist_wheel
 
 # Run
 FROM python:3.6
 
-COPY --from=builder /src/dist/plz_cli-0.1.0-py3-none-any.whl /tmp/
+COPY --from=builder /cli/dist/plz_cli-0.1.0-py3-none-any.whl /tmp/
 RUN pip install /tmp/plz_cli-0.1.0-py3-none-any.whl
 
 ENV PYTHONUNBUFFERED 1
