--- services/controller/src/plz/controller/__init__.py
@@ -1,1 +1,0 @@
-from .controller import Controller  # noqa: F401 (unused)
