+++ services/controller/src/plz/controller/api/__init__.py
@@ -1,0 +1,2 @@
+from .controller import Controller  # noqa: F401 (unused)
+__all__ = [Controller]
