+++ services/controller/src/plz/controller/controller_impl.py
@@ -3,7 +3,6 @@ import logging
 import os
 import random
 import uuid
-from abc import ABC, abstractmethod
 from typing import BinaryIO, Iterator, List, Optional
 
 import requests
@@ -12,117 +11,18 @@ from pyhocon import ConfigTree
 from redis import StrictRedis
 
 from plz.controller import configuration
-from plz.controller.configuration import Dependencies
-from plz.controller.db_storage import DBStorage
-from plz.controller.exceptions import BadInputMetadataException, \
+from plz.controller.api.controller import Controller
+from plz.controller.api.exceptions import BadInputMetadataException, \
     ExecutionAlreadyHarvestedException, ExecutionNotFoundException, \
     InstanceStillRunningException, ResponseHandledException
+from plz.controller.api.types import InputMetadata, JSONString
+from plz.controller.configuration import Dependencies
+from plz.controller.db_storage import DBStorage
 from plz.controller.execution import Executions
 from plz.controller.images import Images
 from plz.controller.input_data import InputDataConfiguration
 from plz.controller.instances.instance_base import Instance, \
     InstanceProvider, NoInstancesFoundException
-from plz.controller.types import InputMetadata
-
-JSONString = str
-
-
-class Controller(ABC):
-    @classmethod
-    @abstractmethod
-    def handle_exception(cls, exception: ResponseHandledException):
-        pass
-
-    @abstractmethod
-    def ping(self) -> dict:
-        pass
-
-    @abstractmethod
-    def run_execution(self, command: [str], snapshot_id: str, parameters: dict,
-                      instance_market_spec: dict, execution_spec: dict,
-                      start_metadata: dict) -> Iterator[dict]:
-        """:raises IncorrectInputIDException:"""
-        pass
-
-    @abstractmethod
-    def rerun_execution(
-            self, user: str, project: str, previous_execution_id: str,
-            instance_market_spec: dict) -> Iterator[dict]:
-        pass
-
-    @abstractmethod
-    def list_executions(self) -> [dict]:
-        pass
-
-    @abstractmethod
-    def get_status(self, execution_id: str) -> dict:
-        pass
-
-    @abstractmethod
-    def get_logs(self, execution_id: str, since: Optional[int]) \
-            -> Iterator[bytes]:
-        pass
-
-    @abstractmethod
-    def get_output_files(self, execution_id: str) -> Iterator[bytes]:
-        pass
-
-    @abstractmethod
-    def get_measures(
-            self, execution_id: str, summary: bool) -> Iterator[JSONString]:
-        pass
-
-    @abstractmethod
-    def delete_execution(self, execution_id: str, fail_if_running: bool,
-                         fail_if_deleted: bool) -> None:
-        """:raises InstanceStillRunningException:
-           :raises ExecutionAlreadyHarvestedException:"""
-        pass
-
-    @abstractmethod
-    def get_history(self, user: str, project: str) -> Iterator[JSONString]:
-        pass
-
-    @abstractmethod
-    def create_snapshot(self, image_metadata: dict, context: BinaryIO) \
-            -> Iterator[JSONString]:
-        pass
-
-    @abstractmethod
-    def put_input(self, input_id: str, input_metadata: InputMetadata,
-                  input_data_stream: BinaryIO) -> None:
-        pass
-
-    @abstractmethod
-    def check_input_data(
-            self, input_id: str, metadata: InputMetadata) -> bool:
-        pass
-
-    @abstractmethod
-    def get_input_id_or_none(self, metadata: InputMetadata) -> Optional[str]:
-        pass
-
-    @abstractmethod
-    def delete_input_data(self, input_id: str):
-        pass
-
-    @abstractmethod
-    def get_user_last_execution_id(self, user: str) -> Optional[str]:
-        pass
-
-    @abstractmethod
-    def kill_instances(
-            self, instance_ids: Optional[List[str]], force_if_not_idle: bool) \
-            -> bool:
-        """:raises ProviderKillingInstancesException:
-
-           :returns bool: false if there are no instances to kill
-        """
-        pass
-
-    @abstractmethod
-    def describe_execution_entrypoint(self, execution_id: str) -> dict:
-        pass
 
 
 class ControllerImpl(Controller):
@@ -146,6 +46,7 @@ class ControllerImpl(Controller):
             self.redis, input_dir=input_dir, temp_data_dir=temp_data_dir)
         self.log = log
 
+    # noinspection PyMethodMayBeStatic
     def ping(self) -> dict:
         # This is plz, and we're up and running
         return {'plz': 'pong'}
