+++ services/controller/src/plz/controller/instances/docker.py
@@ -11,7 +11,7 @@ from plz.controller.containers import ContainerState, Containers
 from plz.controller.images import Images
 from plz.controller.instances.instance_base import ExecutionInfo, Instance, \
     KillingInstanceException, Parameters
-from plz.controller.exceptions import InstanceStillRunningException
+from plz.controller.api.exceptions import InstanceStillRunningException
 from plz.controller.results import ResultsStorage
 from plz.controller.results.results_base import CouldNotGetOutputException
 from plz.controller.volumes import \
