+++ services/controller/src/plz/controller/instances/instance_base.py
@@ -9,7 +9,7 @@ from redis import StrictRedis
 from redis.lock import Lock
 
 from plz.controller.containers import ContainerMissingException, ContainerState
-from plz.controller.exceptions import ProviderKillingInstancesException
+from plz.controller.api.exceptions import ProviderKillingInstancesException
 from plz.controller.results.results_base import InstanceStatus, \
     InstanceStatusFailure, InstanceStatusRunning, InstanceStatusSuccess, \
     Results, ResultsStorage
